auth_system/requirements.txt
fastapi>=0.104.0
sqlalchemy>=2.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
uvicorn>=0.24.0
python-multipart>=0.0.6
slowapi>=0.1.8
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
email-validator>=2.0.0

